+++ apps/src/blockly/cdoBlocklyWrapper.js
@@ -338,9 +338,6 @@ function initializeBlocklyWrapper(blocklyInstance) {
     registerCustomProcedureBlocks() {
       // Google Blockly only. Registers custom blocks for modal function editor.
     },
-    partitionBlocksByType() {
-      // Google Blockly only. Used to load/render certain block types before others.
-    },
     appendSharedFunctions(source, functionsXml) {
       const isXml = stringIsXml(source);
       if (isXml) {
